+++ CHANGELOG.md
@@ -14,6 +14,8 @@ clients compiled against a different minor or major version.
 
 ### Added
 
+- 🚸 Add support for opening existing jobs by their IDs. This feature is
+  optional for device implementations.
 - 👨‍💻 Add stable device IDs and symbol-prefix metadata to exported device targets
   and generated projects ([#475]) ([\@burgholzer])